Separate hadlers for click event from index.js
